+++ tests/infra/consortium.py
@@ -529,13 +529,13 @@ class Consortium:
         return self.vote_using_majority(remote_node, proposal, careful_vote, timeout=30)
 
     def set_js_runtime_options(
-        self, remote_node, max_heap_bytes, max_stack_bytes, max_execution_time
+        self, remote_node, max_heap_bytes, max_stack_bytes, max_execution_time_ms
     ):
         proposal_body, careful_vote = self.make_proposal(
             "set_js_runtime_options",
             max_heap_bytes=max_heap_bytes,
             max_stack_bytes=max_stack_bytes,
-            max_execution_time_ms=max_execution_time,
+            max_execution_time_ms=max_execution_time_ms,
         )
         proposal = self.get_any_active_member().propose(remote_node, proposal_body)
         return self.vote_using_majority(remote_node, proposal, careful_vote)